BJS2-65242 refactoring directories.
