+++ src/main/java/faang/school/postservice/exception/CommentNotFoundException.java
@@ -1,4 +1,4 @@
-package faang.school.postservice.controller.exception;
+package faang.school.postservice.exception;
 
 public class CommentNotFoundException extends RuntimeException {
     public CommentNotFoundException(String message) {
