+++ src/main/java/faang/school/postservice/exception/PostNotFoundException.java
@@ -1,4 +1,4 @@
-package faang.school.postservice.controller.exception;
+package faang.school.postservice.exception;
 
 public class PostNotFoundException extends RuntimeException {
     public PostNotFoundException(String message) {
