+++ src/main/java/faang/school/postservice/service/comment/implementations/CommentServiceImpl.java
@@ -1,10 +1,10 @@
 package faang.school.postservice.service.comment.implementations;
 
 import faang.school.postservice.client.UserServiceClient;
-import faang.school.postservice.controller.exception.AuthorNotFoundException;
-import faang.school.postservice.controller.exception.CommentNotFoundException;
-import faang.school.postservice.controller.exception.PostIdMismatchException;
-import faang.school.postservice.controller.exception.PostNotFoundException;
+import faang.school.postservice.exception.AuthorNotFoundException;
+import faang.school.postservice.exception.CommentNotFoundException;
+import faang.school.postservice.exception.PostIdMismatchException;
+import faang.school.postservice.exception.PostNotFoundException;
 import faang.school.postservice.dto.comment.CommentDto;
 import faang.school.postservice.mapper.comment.CommentMapper;
 import faang.school.postservice.model.Comment;
