+++ src/main/java/faang/school/postservice/service/post/implementations/PostServiceImpl.java
@@ -1,6 +1,6 @@
 package faang.school.postservice.service.post.implementations;
 
-import faang.school.postservice.controller.exception.PostNotFoundException;
+import faang.school.postservice.exception.PostNotFoundException;
 import faang.school.postservice.model.Post;
 import faang.school.postservice.repository.PostRepository;
 import faang.school.postservice.service.post.interfaces.PostService;
