+++ src/test/java/faang/school/postservice/service/controller/CommentControllerTest.java
@@ -3,9 +3,9 @@ package faang.school.postservice.service.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.postservice.config.context.UserContext;
 import faang.school.postservice.controller.comment.CommentController;
-import faang.school.postservice.controller.exception.AuthorNotFoundException;
-import faang.school.postservice.controller.exception.CommentNotFoundException;
-import faang.school.postservice.controller.exception.PostIdMismatchException;
+import faang.school.postservice.exception.AuthorNotFoundException;
+import faang.school.postservice.exception.CommentNotFoundException;
+import faang.school.postservice.exception.PostIdMismatchException;
 import faang.school.postservice.dto.comment.CommentDto;
 import faang.school.postservice.service.comment.interfaces.CommentService;
 import org.junit.jupiter.api.BeforeEach;
