+++ src/test/java/faang/school/postservice/service/post/PostServiceImplTest.java
@@ -1,6 +1,6 @@
 package faang.school.postservice.service.post;
 
-import faang.school.postservice.controller.exception.PostNotFoundException;
+import faang.school.postservice.exception.PostNotFoundException;
 import faang.school.postservice.model.Post;
 import faang.school.postservice.repository.PostRepository;
 import faang.school.postservice.service.post.implementations.PostServiceImpl;
